Add CertStore::export_pkcs12
This exposes the ability to export a `CertStore` as a PKCS#12 archive. Along
the way this also exports the `memory` constructor to allow easily converting a
`CertContext` with a private key to a PKCS#12 archive. Along the way this
removes the `Memory` type and moves the functions to the `CertStore` type
directly.