implement #16: check root certificates against the provided certstore
This adds a check, which compares a root certificate against every certificate
contained by the (optionally) provided certificate store and checks if they
match by comparing their sha256 hashes.
The documentation was updated to reflect that, this therefore
automatically implies trust for certificates added to a passed
certificate store.

Interestingly enough this showed that self-signed.bad-ssl.com
actually updated their certificate
(which still passed the tests with an old implementation during
the course of this PR)